Moved pad function
