--- packages/api-kit/src/utils/data.ts
@@ -1,11 +1,0 @@
-// TODO: remove this function in favor of viem#pad
-export function padHex(
-  hex: string,
-  { dir = 'left', size = 32 }: { dir?: string; size?: number } = {}
-): string {
-  if (size === null) return hex
-  const result = hex.replace('0x', '')
-  if (result.length > size * 2) throw new Error(`Size (${result.length}) exceeds padding size.`)
-
-  return `0x${result[dir === 'right' ? 'padEnd' : 'padStart'](size * 2, '0')}`
-}
