+++ packages/api-kit/src/utils/signDelegate.ts
@@ -1,5 +1,16 @@
 import { Signer } from 'ethers'
-import { padHex } from './data'
+
+// TODO: remove this function in favor of viem#pad
+function padHex(
+  hex: string,
+  { dir = 'left', size = 32 }: { dir?: string; size?: number } = {}
+): string {
+  if (size === null) return hex
+  const result = hex.replace('0x', '')
+  if (result.length > size * 2) throw new Error(`Size (${result.length}) exceeds padding size.`)
+
+  return `0x${result[dir === 'right' ? 'padEnd' : 'padStart'](size * 2, '0')}`
+}
 
 export async function signDelegate(signer: Signer, delegateAddress: string, chainId: string) {
   const domain = {
